Move modules into subdirectory for better file structure.
